fix: modify api versions, and add debug logic

+++ src/client/java/net/hopper4et/oneeyestrongholdfinder/Stronghold.java
@@ -1,11 +1,11 @@
 package net.hopper4et.oneeyestrongholdfinder;
 
 public class Stronghold {
-    long accuracy;
+    double accuracy;
     int x;
     int z;
 
-    public Stronghold(long accuracy, int x, int z) {
+    public Stronghold(double accuracy, int x, int z) {
         this.accuracy = accuracy;
         this.x = x;
         this.z = z;
